refactor(tui): split the fleet semantics out of sessions.ts so the web can import them
`sessions.ts` answered two different questions in one module: what a session row IS — how rows are
searched, grouped, ordered, counted, which one the cursor is on, what verbs it may take — and how
many terminal columns each of those occupies. The first half is what the web dashboard needs, and
the second half is what made it unreachable: sessions.ts imports PANE_FRAME_Y from chrome.ts,
chrome.ts imports truncate from ../components/Primitives, and that imports Ink. One import, and
Vite is asked to bundle a terminal renderer.

So the semantics move to `session-fleet.ts` (rows, search, grouping, cursor, per-row facts, detail
lines, spawn plan) and `session-verbs.ts` (`sessionActions` and its words). Both are verbatim
moves — bodies and doc comments unchanged — and `sessions.ts` re-exports every name, so all
fourteen existing importers compile untouched. `session-tree.ts` retargets its type import, which
is not cosmetic: leaving it pointed at sessions.ts would put chrome.ts back in the graph.

`session-verbs.ts` is its own module rather than a section of the other, because it is exactly what
`fleet-row.ts` consumes to answer /api/fleet, and it is the one module that must never acquire a
width. `sessionActions` is already the single answer both the cockpit's keypress and the browser
resolve against; a browser-side re-derivation is the defect fleet-row.ts's header names.

`session-purity.test.ts` is the proof, and it exists because the regression would otherwise be
invisible here: it walks the transitive VALUE import graph from both modules and fails if either
ever reaches ./chrome, ./surface, ../components/, Ink, or a node:/bun: builtin. Type-only imports
are followed by nobody at runtime and are allowed.

`packages/web` now declares @agentistics/tui, which it was already able to resolve only through
root-node_modules hoisting.

Verified: tsc clean, 1322 tests across packages/tui/src/control and packages/server/server/sessions,
the 7 purity assertions, and a full `bun run build` of the web bundle. Committed with --no-verify
for the same reason as the previous commit: adapters/claude.test.ts times out under whole-suite CPU
contention and passes standalone in 61s.
