+++ packages/tui/package.json
@@ -10,6 +10,8 @@
     "./control/altScreen": "./src/control/altScreen.ts",
     "./control/stream": "./src/control/stream.ts",
     "./control/sessions": "./src/control/sessions.ts",
+    "./control/session-fleet": "./src/control/session-fleet.ts",
+    "./control/session-verbs": "./src/control/session-verbs.ts",
     "./control/search-scope": "./src/control/search-scope.ts",
     "./control/surface": "./src/control/surface.ts",
     "./control/i18n": "./src/control/i18n.ts"
