+++ packages/web/package.json
@@ -10,6 +10,7 @@
   },
   "dependencies": {
     "@agentistics/core": "workspace:*",
+    "@agentistics/tui": "workspace:*",
     "@opentelemetry/api": "^1.9.1",
     "@xterm/xterm": "^5.5.0",
     "date-fns": "^4.1.0",
